1.0.3
Added image.
Skitarii Helmets now give low oxygen warning when oxygen is low and in water.
Added Pomegrenade Extract as phantom item to help with mod compatability. It is not obtainable in normal gameplay and if spawned in with console commands will do nothing, it just exists in the code.
Removed throwable tags from a bunch of magazines.
Fixed Assistant Talents to work properly and give correct buff icons.